feat: allow to define whether link should open in new window
